Débug V3.2 terminé (test passe)

+++ apps/plant_shop_rust/plant_shop_rust_see_orm/migrations/init.sql
@@ -15,7 +15,7 @@ CREATE TABLE plants (
 	id SERIAL PRIMARY KEY,
 	name VARCHAR(100) NOT NULL,
 	description TEXT,
-	price NUMERIC(10,2) NOT NULL,
+	price INTEGER NOT NULL,
 	stock INTEGER NOT NULL DEFAULT 0,
 	created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
@@ -23,7 +23,7 @@ CREATE TABLE plants (
 CREATE TABLE orders (
 	id SERIAL PRIMARY KEY,
 	user_id INTEGER REFERENCES users(id) ON DELETE SET NULL, -- ON DELETE SET NULL pour garder l'historique
-	total NUMERIC(10,2) NOT NULL,
+	total INTEGER NOT NULL,
 	status VARCHAR(50) NOT NULL DEFAULT 'pending', -- Ajout de la colonne status
 	created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
@@ -33,7 +33,7 @@ CREATE TABLE order_items (
 	order_id INTEGER REFERENCES orders(id) ON DELETE CASCADE,
 	plant_id INTEGER REFERENCES plants(id) ON DELETE SET NULL, -- ON DELETE SET NULL pour ne pas perdre l'item si la plante est supprimée
 	quantity INTEGER NOT NULL,
-	price NUMERIC(10,2) NOT NULL
+	price INTEGER NOT NULL
 );
 
 -- Index et contraintes utiles
