+++ apps/plant_shop_rust/plant_shop_rust_see_orm/src/order_items/models.rs
@@ -1,5 +1,4 @@
 use serde::{Serialize, Deserialize};
-use sea_orm::prelude::Decimal;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct OrderItem {
@@ -9,8 +8,7 @@ pub struct OrderItem {
     pub order_id: Option<i32>,
     pub plant_id: Option<i32>,
     pub quantity: i32,
-		#[serde(serialize_with = "crate::plants::models::serialize_decimal_as_i32")]
-		pub price: Decimal,
+		pub price: i32,
 }
 
 #[derive(Deserialize)]
@@ -18,5 +16,5 @@ pub struct NewOrderItem {
     pub order_id: i32,
     pub plant_id: i32,
     pub quantity: i32,
-    pub price: Decimal,
+    pub price: i32,
 }
